+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unpublished Changes
+* Add navigation and navigation-item modules from theme hubs work.
+
 
 ## 1.72.0 06-06-2025
 * larva-js - Play JWPlayer with muted audio by default.